订单流程 （待付款，取消子单，取消整单。已付款，退子单，退整单《待出行，未出行，已出行》。退款单。）

+++ LetsGoApp/app/src/main/java/com/njz/letsgoapp/bean/order/OrderRefundDetailChildModel.java
@@ -43,7 +43,7 @@ public class OrderRefundDetailChildModel {
     private String title;
     private int roomNum;
     private String travelDate;
-    private int payPrice;
+    private float payPrice;
     private float price;
     private int ticketNum;
     private int dayNum;
@@ -142,11 +142,11 @@ public class OrderRefundDetailChildModel {
         this.travelDate = travelDate;
     }
 
-    public int getPayPrice() {
+    public float getPayPrice() {
         return payPrice;
     }
 
-    public void setPayPrice(int payPrice) {
+    public void setPayPrice(float payPrice) {
         this.payPrice = payPrice;
     }
